Fixed ambiguous pow errors

+++ include/updownpass.h
@@ -23,7 +23,7 @@ class UpDownPass : public Kernel, public Logger {
     real_t x = 1.0 / THETA;                                     // Inverse of theta
 #if ERROR_OPT
     assert(THETA != 1.0);
-    real_t a = c * std::pow(std::abs(C->M[0]),1.0/3);           // Cell coefficient
+    real_t a = c * powf(std::abs(C->M[0]),1.0/3);               // Cell coefficient
     for (int i=0; i<5; i++) {                                   // Newton-Rhapson iteration
       real_t f = x * x - 2 * x + 1 - a * std::pow(x,-P);        //  Function value
       real_t df = (P + 2) * x - 2 * (P + 1) + P / x;            //  Function derivative value
@@ -68,7 +68,7 @@ class UpDownPass : public Kernel, public Logger {
     startTimer("Upward pass");                                  // Start timer
     C_iter C0 = cells.begin();                                  // Set iterator of target root cell
     postOrderTraversal(C0, C0);                                 // Recursive call for upward pass
-    real_t c = (1 - THETA) * (1 - THETA) / std::pow(THETA,P+2) / std::pow(std::abs(C0->M[0]),1.0/3); // Root coefficient
+    real_t c = (1 - THETA) * (1 - THETA) / std::pow(THETA,P+2) / powf(std::abs(C0->M[0]),1.0/3); // Root coefficient
     setRcrit(C0, C0, c);                                        // Error optimization of Rcrit
     if( cells.size() > 9 ) {                                    // If tree has more than 2 levels
       for (C_iter C=C0; C!=C0+9; C++) {                         //  Loop over top 2 levels of cells
